Made the property "id" serialize into a JSON object to send.
"id" did not serialize because of this contract resolver method.
"id" was made not to serialize before because "id" would be extracted
from the object and put into the request url. Now, due to operations
such as bulk-insert operations, "id" properties are required to be
present in the JSON object request as well as in the url.