ContractResolver resolves when to serialize ID and when to not (SEE DESCRIPTION)
Updating Rows requires IDs the in json object so we must make ID
serializable in this situtaion. Howerver I have not find a clean cut way
to differentiate between knowing whether the Row object being serialized
originates from the UpdateRows or the AddRows method. So what I did is I
put a rule to ALWAYS serialize ID for 'Row' types and then make all of
the row IDs in AddRows to null.