APISDK-1596,1748,1740,1352,1410
expose changeAgent, assumedUser, userAgent on Smartsheet client
remove apiScenario from the builder and constructors as it is only meant to be used internally. Allow for RestSharp request creation to be overridden for the injection of additional headers or proxy information. Implement Api-Scenario as an extended DefaultHttpClient class
changed method for providing custom ShouldRetry. This now uses the same method of extending DefaultHttpClient and overriding ShouldRetry and CalcBackoff.